fix: middleware in array to route generic validation
