+++ app/routes.ts
@@ -9,8 +9,7 @@ import { userValidation } from './middlewares/user-validation';
 export const init = (app: Application): void => {
   app.get('/health', healthCheck);
   app.get('/users', getUsers);
-  app.post('/users', userValidation);
-  app.post('/users', createUser);
+  app.post('/users', [userValidation, createUser]);
   app.get('/users/:id', getUserById);
   app.get('/todos', getTodos);
   app.get('/info', getInfo);
